Add support for non-pg_catalog codecs
Not all data types have predefined OIDs, but we still want to provide
efficient codecs for more popular extension types, such as hstore.

The only difference between "builtin" and "extra" codecs is that the
latter are registered using a unique name rather than an OID.  In order
to enable {en,de}coding of the desired type, an application must call
Connection.alias_type to specify the name and the schema of the type and
the unique name identifying the extra codec.

For example:

    conn = await asyncpg.connect()
    await conn.alias_type('hstore', schema='public',
                          alias_to='pg_contrib.hstore')

This commit also includes the implementation of the hstore codec.