Add infrastructure for multiformat data codecs
Current asyncpg codec architecture allows for only one codec per data
type.  This arrangement works most of the time, but fails utterly to
handle situations when the text format I/O is forced by the lack of
binary I/O support for one of the types in the composite type.

This commit is the preparation for the introduction of text format
parsing and allows to associate multiple format-specific codecs for each
type.