Implement conistent handling of timeouts.
The timeout logic is currently a bit of a mess.  This commit attempts to
tidy things up.

Most importantly, the timeout budget is now applied consistently to the
_entire_ call, whereas previously multiple consecutive operations used
the same timeout value, making it possible for the overall run time to
exceed the timeout.

Secondly, tighten the validation for timeouts: booleans are not
accepted, and neither any value that cannot be converted to float.