Implement an Interval class and use that instead of datetime.timedelta
This is a POC, targeting issue #150: PostgreSQL interval data type and Python timedelta have a
similar goal, but are subtly different because PG date arithmetic is much more versatile.

So, instead of trying to coerce an interval to a timedelta, the new class aims to carry the
same bits of information, and may implement a similar semantic, if needed.