Allow aliasing builtin types by name in set_builtin_type_codec()
Currently, `Connection.set_builtin_type_codec()` accepts either a
"contrib" codec name, such as "pg_contrib.hstore", or an OID of
a core type.  The latter is undocumented and not very useful.

Make `set_builtin_type_codec()` accept any core type name as the
"codec_name" argument.  Generally, the name of the core type can be
found in the pg_types PostgreSQL system catalog.  SQL standard names
for certain types are also accepted (such as "smallint" or
"timestamp with timezone").  This may be useful for extension types
or user-defined types which are wire-compatible with the target
builtin type.