Compatibility with pgbouncer with pool_mode != session
Add `session` parameter to `connection()` with default value `True`.
When set to `False` the client will limit usage of prepared statements
to unnamed in a transaction only.

As a side effect handling of unnamed prepared statements improved
irrespective of `session` value - they are automatically reprepared
if something invalidating them occurs.