Stop treating ReadyForQuery as a universal result indicator
ReadyForQuery is special in auth and simple query flows, where it _does_
signal the final confirmation of the result, but in all other flows
other, more specific messages do that.  Now, asyncpg will use the
rules of a particular subprotocol when determining the timing of the
result waiter wakeup.  These changes also make most cases of Sync
emission unnecessary, although that cleanup will be addressed in
subsequent commits.  This consolidation also results in a nice
reduction of duplicated code.