Update CI matrix, aarch64 builds (#595)
This does the following changes:

* Updates the Python versions for the OSX builds to the latest ones
* Removes the version pin on wheel
* setuptools 48.0.0 broke asyncpg, distutils is deprecated now, I've rewrote it to setuptools entirely
* CI pipelines now have names
* aarch64 tests against the latest python and postgresql are now done
* manylinux2014 wheels for aarch64 are also built now
* bumped Ubuntu to focal (20.04)
* bumped Cython to fix compatibility with python 3.10
* CI builds now actually install the built wheel and not the latest pypi release
* version test is skipped for dev builds as the git commit hash mixes up the version number
* resolved all travis config warnings
* postgres 9.2-9.4 are EOL and removed from the tests

The manylinux containers from pypa only work on the architecture they are made for, so a aarch64 manylinux wheel can only be built on aarch64. Therefore, two containers now perform the builds. It might be worth thinking about letting the aarch64 container handle the upload alone for all wheels as it is ages slower than the x86 one anyways.

I've tested pretty much everything except the S3 bucket and pypi push, the wheels themselves get built and the other pipelines pass.