Don't build aarch64 wheels for now
The porting of arm64 builds to Github Actions uncovered an
architecture-specific (or, perhaps, virtualization-specific) bug in
asyncpg, so, rather than blocking the release, drop aarch64 wheels for
now.  Also, `manylinux2014_aarch64` is not considered stable yet
(see pypa/manylinux#84) and so the compiled wheels might not even work
correctly on all aarch64 boxes.