Make Server Version Extraction More Flexible
https://github.com/MagicStack/asyncpg/issues/250
https://github.com/MagicStack/asyncpg/issues/261
https://github.com/MagicStack/asyncpg/issues/771

Just tried to use this library through Ormar->Databases->AsyncPG against
a Yugabyte cluster and hit issue 771. Looks like this has been a problem
for a while now so going for a complete overhaul of the server version
extraction method. Using a groupdict regex against the version string
allows for much higher flexibility in extracting what we're looking for
and fixes 771 while not breaking any of the existing version patterns.