add quote param to add_listener
This param is a switch that lets the user specify whether or not they
want the channel name to be sent to postgres as quoted identifier. If
the flag is set to false, it allows the user to call the method with
all-caps channel names without the need to call the channels as quoted
identifiers from notify statements in postgres. The flag defaults to
True so will not break anyone's code.