Allow to specify custom connection reset query builder
The default connection reset behaviour might be an overkill
for some cases e.g. when respective server capabilities are not used
or resources are cleaned up explicitly before returning the connection
to the pool.

This provides an opt-in way to override this default behaviour.

Also see https://github.com/MagicStack/asyncpg/issues/780