Make `prepare()` not use named statements by default when cache is disabled
We allow disabling the statement cache to circumvent a proxy's inability to
cope with them, however there are still some holes where the use of a named
statement is attempted: `prepare()` with the default arguments, and
`copy_in()`.  Fix both so that the use of a named statement is dependent
upon whether the prepared statement cache is enabled.

Fixes: #1219
Closes: #1218