Avoid leaking connections if _can_use_connection fails (#1269)
If _can_use_connection fails (say, because of an asyncio timeout), then we may have a full connection that has been created but would be leaked by this function. (note: `_connect_addr` takes care of cleaning up after itself if it fails partway)

This is particularly possible in the case of pgbouncer, where we may succeed at establishing a connection much quicker than even a trivial call to the backing database would take.

I believe this failure mode was introduced in https://github.com/MagicStack/asyncpg/pull/987

I'm not certain if we should `await` here or just punt the closing of the other connections to a background task (I don't know how risky `close` is on an established connection).